widget: rename Child to GetChild
Widgets with children need both a field for the child and a method to
get the child. Because users create widgets by using struct literals and
because users will more often instantiate widgets than call the method
for getting the child, we should optimize for having the field named
Child.