widget: emulate inheritance with embedding and self-reference
We implement inheritance via the following pattern:

	type Abstract interface {
		Foo()
		Bar()
	}

	type Concrete1 struct {
		Self Abstract
	}

	func (c Concrete1) Foo() { c.Self.Bar() }
	func (c Concrete1) Bar() { println("Concrete1.bar") }

	type Concrete2 struct {
		Concrete1
	}

	func (c Concrete2) Bar() { println("Concrete2.bar") }

This allows custom element implementations to embed the appropriate base
type (e.g. ElementMixin or RenderObjectElementMixin) and only override
the methods that need custom behavior. This requires less boilerplate
and is easier to document than having to define methods that act as
wrappers around default implementations.

The approach is not without drawbacks, however. Any type that partakes
in this fake inheritance has to be careful to always call methods on
Self, not the receiver itself. Similarly, functions that accept Element
arguments have to be passed Self.