widget: don't emulate inheritance
The drawbacks of the fake inheritance approach outweighed the benefits.
We switch to a new API that decomposes the Element interface into many
tiny, optional interfaces and that provides interfaces that help with
implementing specific kinds of elements.

This works best in combination with a tool like impl (or the gopls
equivalent), as the tool will quickly generate all the methods that the
user has to define to call the provided default implementations.

This is a middle ground between requiring every implementation to
implement all methods, and automatically providing default
implementations for everything.