opentype: add (incomplete) package for parsing OpenType files
We began work on this before we decided to switch to the C++ version of
Harfbuzz, which comes with its own OpenType functions. This package was
supposed to replace the parser used by go-text's Harfbuzz port. Unlike
their parser, ours doesn't have to load files into memory and instead
works with mmaped files and lazy parsing.

Even though we won't end up using this package for shaping, we still
keep it around in case we want to implement font databases or tools for
debugging fonts.