internal/sparse: add optimization pass for wide tile commands
Some frontends, like our Lottie renderer, create a large number of
nested layers, many of which are irrelevant to the final result.
Furthermore, there is a lot of overdraw that is trivial to detect and
avoid.

While the optimization pass accounts for 10% of total CPU time (for
Lottie animations), which is quite a lot, it reduces overall CPU time
spent in fine rasterization by a factor of 10, making it well worth it.
A significant chunk of the cost of the optimization pass is in
maintaining 256-bit wide bitmasks, which could be sped up easily with
SIMD.